Split HA server command tests from reload tests
